requirements_fm.txt
distance==0.1.3
einops==0.8.1
h5py==3.11.0
matplotlib==3.8.2
metric==0.10.0
munch==4.0.0
numpy==2.2.4
opencv_python==4.11.0.86
Pillow==11.1.0
PyYAML==6.0.1
PyYAML==6.0.2
tensorflow==2.16.1
torch==2.6.0
torchdiffeq==0.2.5
torchdyn==1.0.6
torchvision==0.21.0
tqdm==4.66.2
transformers==4.40.1
